Avoid unnecessary traversals of security namespace mount list.
Testing indicated that the traversal of the namespace mount list
was a hot list operation, given that it is being called on every
path lookup.  The traversal of the mount list does not need to be
conducted if the path was previously looked up and found to not
be a mountpoint.

In order to optimize this process, a tsem_inode_instance
structure was added to the tsem_inode structure to cache the
mount status of the inode.  When an inode is instantiated the
tsem_inode_instance pointer is initialized to an error pointer
value of -ENOENT to indicate that the mount status of the inode
has never been determined.

The _fill_mount_path() function was updated to set the status of
of the tsem_inode_instance structure depending on the result of
the lookup of the pathname on the mount list.  If the pathname
was found on the mount list the tsem_inode_instance pointer is
set to the tsem_inode structure that labels the inode.

If the pathname was not found on the mount list the pointer is
set to a NULL value to indicate that there is no subsequent need
to iterate over the mount list for that pathname.

The fill_path() function was updated to test the status of the
tsem_inode_instance pointer that is used to cache the mount
status of the pathname.  If the tsem_inode_instance pointer is
valid the contents of the structure are used to initialize the
path instance elements of the path description structure.

The get_sb_mount() function was updated to traverse the mount
list if no mount device was specified.  This catches instance
information about changes in the mount status of the path.