Release mount description on call to sb_umount.
A combination of testing and code review indicated the need to
remove a mount description from the security modeling namespace
mount list when the tsem_sb_umount event handler is called.

Search the security modeling namespace mount list when the event
description for the unmount description is generated and free the
entry that matches the pathname being unmounted.