Implement early event termination flag.
This commit begins the process of generalizing TSEM so that
arbitrary security models can be implemented with loadable
modules that provide implementations for the various modeling
methods.

In order to properly support the generic handling of security
events support is needed to terminate event handling at arbitrary
points in the event processing chain.

Add a boolean variable to the tsem_event structure that can be
set by a method implementation to indicate that further
processing of the event is not required.