Convert tsem_task_alloc to standard event handling.
In order to allow the tsem_task_alloc security handler to be
treated by a loadable model it needs to move from a direct inline
implementation to the standard TSEM event handling implementation
that acts on a security event description structure.

The following structure was added to the tsem.h definitions:

tsem_task_args

This structure will be used to convey arguments regarding the
task that is being initiated or freed to the event handlers.

These events that were previously handled by inline
implementations in the security event handlers will be classified
as 'internal' implementations since they will not be modeled by
either internal or external trusted modeling agent
implementations.

The enumeration ordinals for these events are added after the
TSEM_EVENT_CNT ordinal that ends the enumeration of events that
are modeled by the TSEM implementation.

The 'terminate_event' flag that was added to the tsem_event
structure is tested after the call to the tsem_event_init()
function since the implementation of the handler for this event
will be executed as part of the event initialization.  The
event handler code sets this flag to true after running the task
allocation initialization.

Since the tsem_task_alloc() function may be implemented before
the TSEM infrastructure is fully initialized the event is
captured at the start of the tsem_event_init() function in order
to allow these events to be processed before the TSEM is fully
initialized for modeling.  This is consistent with the behavior
of the handler when it was implemented directly in the
tsem_task_alloc() function.

The initialization of the new task was implemented in the
event.c:task_alloc() function and is a direct replication of the
code that was formerly directly processed.

Moving forward the plan will be to have the event initialization
function be a method that can be configured by an alternate
method supplied by a loadable module.  That will allow that
method to implement a customized implementation of the task_alloc
functionality.