Always transfer namespace context to new task.
Moving the functionality of the tsem_task_alloc security event
handler to the event processing code resulted in what was a
difficult to locate regression when support for overriding the
default event initialization function was added.

The security modeling namespace context information always needs
to be transferred from the previous task to a new task being
started in a security modeling namespace regardless of the model
being implemented in the namespace.  If that is not done models
that wish to override the default event initialization will
generate segmentation faults when the context information is
referenced since without the transfer the context pointer is null.