Use default system event initialization.
The implementation of system event implementation as a method for
a security modeling namespace was not protecting against a
scenario where the default event initialization routine,
event.c:tsem_event_init() was desired.

Check to see if an alternate method has been specified and if not
call the default initialization function.