Treat all event enumerations in the same manner.
When the implementation of the tsem_task_alloc handler was moved
into the general event processing code the enumeration type
included the string _INTERNAL_ in the enumeration description.
There was also not an entry in the tsem event names array for
that event.

That turned out to be a poor choice and generated an
implementation inconsistency that was unneeded.

Remove the _INTERNAL_ designation in the enumeration type and add
a description of the event name for the purposes of consistency.