Convert bprm_committed_creds handler to standard event processing.
In order to further the generalization of the TSEM event handling
infrastructure the implementation of the tsem_bprm_committed_creds
handler for the TSEM model was converted to standard event
processing.

The bprm_committed_creds() function was added to the event.c file to
encapsulate the functionality that had previously been in the
tsem_bprm_committed_creds() function.

A switch statement was added to the tsem_event_init() function to
dispatch the handling of functions that were previously
implemented in the security event handlers themselves, in
anticipation of moving other functionality into the event handler
for a specific model.

Since the task instance count was being incremented in the
tsem_bprm_committed_creds() function the 64 bit atomic counter
was moved from the tsem.c file to the event.c file since that is
where it was being referenced.

These changes allowed the map.c:tsem_map_task() function to be
simplified.