Implement a model initialization method.
Moving forward one of the issues that needs to be addressed is to
provide a method for a namespace specific model to be
initialized.

Add a model_init member to the tsem_context_ops structure to
allow a method to be specified to initialize a model that is
attached to a security modeling namespace.

For the internal TSEM model the model initialization consists of
adding the aggregate value to the model.

Change the following function name:

tsem_model_add_aggregate()

To

tsem_model_init()

In order to align the name of the function to its generic
purpose.