Handle reference for TSEM context in tsem_task_alloc.
A reference count regression was noted when testing was conducted
with the null TSEM model.

This was caused by the fact that the reference to the TSEM
context structure for security modeling namespace was being
handled in the TSEM model specific event initialization code.

Since the tsem_task_free() function was directly calling the
tsem_ns_put() function regardless of the model being used the
result was a reference count underflow.

Fix this problem by moving the kref_get() function call for the
context reference count into the tsem_task_alloc() function.
This ensures that the reference count generated by every task
allocation is matched with a reference decrement when the task is
retired.