Add support for the security_bprm_check_security event.
A number of additional event handlers are needed in order to
support the ability to use Tomoyo as a demonstration of the
ability for TSEM to implement arbitrary models.

This commit implements the tsem_bprm_check_security event handler
to support this new event type.

The TSEM model does not use this event handler so this commit
introduces the pattern of using C99 array initializers to
implement specific event bypasses.

The only argument taken for this event handler is a pointer to a
linux_binprm structure.  As a result a simple pointer is added to
the CELL definition union since there is no justification for
retention values for the argument.