Implement handler for security_cred_prepare.
The tsem_cred_prepare() function was added to handle these
events.

The tsem_cred_prepare_args structure was added to encapsulate the
arguments passed to the handler.  Since the TSEM model is not
handling these events a retention structure was not implemented.

A bypass was added for the model0 implementation.