Implement handler for security_file_truncate event.
The tsem_file_truncate() function was added as a drive for this
event handler.

Since this handler takes a pointer to a file structure as an
argument the tsem_file_args structure was used to carry the
arguments to the model implementations.

A bypass was added for the TSEM model0 model.