Implement handler for security_file_ioctl_compat.
Tomoyo provides a handler for this security event so support was
added to allow TSEM to handle this event.

The handling of this event is very similar to the
security_file_ioctl event that TSEM already has a handler for.
The infrastructure from that event was largely re-used for this
event.

An event bypass was added for the TSEM model0 implementation.