Implement handler for security_path_chroot.
This commit represents implementing support for the last path
based security handler that is needed by the Tomoyo LSM.  We are
using this LSM as an example that TSEM can implement any current
and existing LSM in comparmentalized security modeling namespace.

The tsem_security_path function was implemented in order to
aggregate the input argument which is a simple pointer to a path
structure.

The TSEM model0 implementation was exempted from handling this
event.