Free task private memory.
A void pointer was previously added to the TSEM task structure to
allow models to store model specific task information.

The expectation is that a model will use a single memory
allocation to implement the model specific task information.  On
the retirement of the task unconditionally free the contents of
that pointer.  If no model specific information was implemented this
pointer will have a NULL pointer that will be a no-op for the
freeing function.