Make vfsmount structure pointer available.
Previously the tsem_sb_umount() event handler had populated the
dentry pointer in the 'in' union of the tsem_sb_args structure
from the vfsmount structure pointer that had been passed to the
handler.

In order to support alternate models that may want access to the
vfsmount structure add a pointer for this structure to the 'in'
union of the structure.

The event.c:get_sb_umount() function had used the dentry
structure pointer from the 'in' member of the structure.  Modify
that function to obtain the dentry pointer from the vfsmount
structure pointer.