Allow struct sock and struct socket pointers to be dispatched.
In order to increase the flexibility of the models that can be
implemented the tsem_socket_args structure was modified in order
to allow either a struct sock pointer or a struct socket pointer
to be passed to the event implementations.  The original TSEM
only model had marshalled and retained arguements based on a
struct sock implementation.

Two anonymous unions were added to the tsem_socket_args structure
to allow either a struct sock or struct socket pointer to be
stored in the location where a struct sock pointer had been
originally used.

The event handlers that were previously extracting a struct sock
pointer from the struct socket argumenet were modified to place
the struct socket pointer into the 'in' union of the socket.

The event marshaling code in the event.c file was modified to
extract the struct sock pointer from the socket pointer in order
to marshall the retention form of the event arguments.