Update exported parameters for socket send and receive messages.
In order to better support alternate security models the
following event handlers were updated:

tsem_socket_sendmsg
tsem_socket_recvmsg

Following the model for previous socket based event handlers the
primary event handlers were updated to pass a pointer to the
socket argument rather than the struct sock pointer that was
inside the socket structure.

The addr member of the 'in' union in the tsem_socket_args
structure was enclosed in an anonymous union that now includes a
pointer to a msghdr structure.  This provides a means for passing
a description of the message to event handlers.

The code that was abstracting information for the TSEM model0 was
moved to the event.c:get_socket_msg() function.  That function
was updated to extract the struct sock pointer from the socket
argument that was in the input marshalling structure.