Implement framework for path based security events.
This commit begins the process of implementing event handlers for
events that are generated by the path based security events.
These events are generated when the CONFIG_SECURITY_PATH
configuration variable is set for a kernel build.

These events are similar to the inode base events but are called
with a pointer to a path structure rather than an inode.  In
addition to Tomoyo these event handlers are used to drive the
AppArmor security framework.

Two sections of code were added to the tsem.c file that are
protected by a test for the CONFIG_SECURITH_PATH CPP define.  The
first section contains the implementations of path based security
event handlers while the second session guards the definitions of
the event handlers.

The path based framework is started with a handler for the
security_path_truncate event.  This handler and subsequent
handlers are very similar in implementation to the inode based
equivalent functions with the exception that a pointer to a path
structure is included in the CELL definition.  Since the TSEM
model0 implementation does not use these handlers the path
structure pointer is not included in the 'in' union nor are there
retention versions of the path retained.

Since most of these events are very similar to their inode
variants the path pointer was added to the tsem_inode_args
structure of the CELL definition.