Implement dispatching of the security_task_free event.
In order to allow models to free any memory that it may have
allocated for private task data the TSEM event handler code was
modified to dispatch the security_task_free event through the
standard event handling chain.  This allows the event to
intercepted by the model specific initialization code.

This event handler needed to be modified to bypass the standard
event allocation code and use an event description structure that
is allocated on the stack.  The freeing of a task structure is
deferred to a workqueue context from which the security_task_free
handler is called.  This ends up causing a lock conflict when the
event description structure is freed.

Since the TSEM model0 does not require any special freeing
procedures for the task information the event is intercepted by
the switch statement in the tsem_event_init() function that
handles events that TSEM is not retaining.