Change from bypassing to accepting events.
Previously the tsem.c:bypass_event() function treated the
.bypasses member of the tsem_context_ops structure as a boolean
array of events that the model was not handling.

Experience with implementing TOMOYO as a loadable model plugin
suggested that it was more logical to specify the events that
would be handled rather than bypassed.

The .bypasses member of the function was changed to .events in
order to indicate that this member specifies the events to be
processed.

The bypass_event() function was modified to invert the test so
that a true value was returned if the entry for the handler was
not set to a true value.

The event specification array for TSEM model0 was modified to
specify all of the events that the model was implementing
handlers for.