Avoid indirect function call for TSEM model0.
The ability to implement loadable modules required the
implementation of indirect function calls in order to allow the
event processing to be customized based on the needs of the model
being implemented.

Indirect calls are considered problematic secondary to the need
to mitigate speculative execution vulnerabilities.  These
mitigations introduce what are considered to be significant
performance regressions and there are ongoing efforts to mitigate
these issues.

Static calls have been introduced in an attempt to mitigate this
issue but the use of these are inconsistent with a base class
model where methods are overriden, a model that TSEM is
implementing.

As a preliminary step in mitigating criticisms over indirect
function calls the points in the code where the event handling
methods are implemented were converted to a model that detects
whether or not a method override has been specified.  If a method
override has not been provided a direct call is made to the TSEM
model0 implementation.

The check for a NULL override method was wrapped in a likely()
function in order to optimize for this case.