Strengthen correspondence between TSEM types and names.
One of the criticisms that Casey Schaufler had raised was over
the fact that it would be difficult to avoid incompatibilities
between the tsem_event_type enumeration and the tsem_names array.
He noted that it would be inevitable for a change to be made in
one that was not reflected in the other.

He also noted that he did not have a good solution for the issue.

This commit implements what is believed to be a fix to the issue
that will prevent a discordance from developing between the
enumeration type and the names array.

A BUILD_BUG_ON call was added to the tsem.c:tsem_init() function
where it will always be called.  The conditional checks to see
whether or not the size of the tsem_names array is consistent
with the number of enumeration types that have been defined.  If
this is not the case a build error is generated.

The following tsem_event_type enumeration was added in order to
have a direct relationship between the enumeration types and
strings in the tsem_names array:

TSEM_UNDEFINED

This matches the 'undefined' string that is the first element in
the tsem_names array.

The explicit sizing of the tsem_names array was removed so that
the size of the array will be defined by the number of elements
declared in the array.

The specification of the strings for the array was changed from
the former standard initialization to named declarations for each
element in the array using the tsem_event_type enumerations.

If an enumeration type is added without a corresponding member
being added to the array a build error will be generated.   The
inverse is also true, if a member of the array is declared
without an enumeration type being declared the build will break.

The use of named array member declarations will ensure that the
enumeration type always matches its name.