Protect string length check of mount device name.
Testing on a Debian distribution demonstrated that the system was
panicking on boot.

The root cause of the problem was easily demonstrated and was
secondary to the fact that the dev_name argument to the
security_sb_mount event handler was a NULL in some instances.
Since the action by the system was to activate the autofs mount
system it is presumed that the security event handler is being
called by that infrastructure with a NULL pointer.

Implement a dual check conditional that first evaluates for a
null dev_name pointer value before calling the strlen() function
on that pointer.