Update handling of integrity file opens.
A regression was noted when testing with a 6.1.x kernel
implementation on Debian.  The systemd journal infrastructure was
not starting, secondary to an error being generated by the
security_file_fcntl event handler.

After an extensive investigation the problem was ultimately traced
to calls involving event handlers using file descriptors that
were opened with the O_PATH directive.  The O_PATH directive is
designed to return a file handle with limited functionality that
can be used, for example in issueing fcntl calls against the
file.

A tip to finding this issue was a patch to the IMA infrastructure
from Roberto Sassu in which any file opens with the O_PATH
directive were bypassed from integrity checking.  This was under
the assumption that since the file cannot be read or written
there was no need to have the file open fall under the integrity
protection envelope.

While that issue is relevant to IMA there is a need by TSEM to
obtain the cryptographic checksum of any file that is opened
since the task identity model is dependent on the security
coefficient generated by an event being tied to the identity of
the file.

In the process of evaluating how TSEM uses the kernel read
infrastructure to open files it was noted that the scheme that
TSEM uses, adopted from earlier IMA implementations, needed to be
updated.

The IMA strategy now opens a new file descriptor if the file is
deemed not to have any type of read operations available to it.
That implementation does not indulge in the previous technique of
modifying the file flags with restoration of the flags after the
I/O completes.

The event.c:open_event_file() function was updated to follow this
strategy.  The f_mode value of the file descriptor must have the
FMODE_READ and FMODE_CAN_READ bits be set in the mode
description.

The key to fixing the errors that were being generated by TSEM
was to include the O_PATH directive in the flags that were to be
masked out in the new file description.

Previously a pointer to an unsigned integer was passed to the
open_event_file() function in order to allow that function to
signal what issues needed to be unwound after the kernel read
was completed.  This argument was converted to a pointer to a
boolean value since the only signal that needs to be indicated is
whether or not a new file descriptor instance was opened.

Since the Debian implementation enables IMA an additional
regression was addressed.

The IMA implementation increments the number of read instances
that are open for an inode.  Putting the file descriptor
decrements this count.  The IMA implementation tests for an
underrun of this counter and generates a BUG() if it does so.

In order to address this issue an inline function was implemented
in the event.c file that is conditionally compiled to a function
that increments this count if IMA is enable or turns into an
empty function implementation if IMA is not enabled in the
kernel.