Narrow locking scope for file digest cache lists.
Testing with the version 1.2.x release of runc, that was using
the overlay filesystem as part of its isolation strategy, was
generating lockdep violations secondary to a clash between the
inode lock and the mutex lock that protects access to the list of
digests attached to the inode.

The strategy embraced by the IMA infrastructure was to define new
lockdep classes for the inode integrity mutex lock.

Rather than implement that level of complexity, it was apparent
from code inspection, that the window over which the digest lock
was being held could be significantly reduced.

The digest mutex lock that was being held during most of the
duration of the event.c:add_file_digest() function was removed
and substituted with the application of the digest mutex lock
over critical sections in the event.c:find_digest() and
event.c:add_digest() functions.

This change addresses the lockdep issue and significantly narrows
the time period which the digest mutex lock needs to be held.