Fill dentry path only if path device is found.
The event.c:fill_path_dentry() function was inconsistent with its
path based partner event.c:fill_path() in that it was requesting
the generation of the dentry based path regardless of whether or
not the get_root() function had set a device type.

Conditionalize the generation of the path based on the status of
the dev member of the path structure.