Optimize return processing in fill_path().
A review of the event.c:fill_path() function suggested that the
handling of error returns could be simplified and exit via the
done: label could be eliminated.

Return any error based results immediately rather then deferring
the return of the error to the end of the function.  This method
was preferred since there are no locks or other state information
that needs to be cleared.