Simplify and deconditionalize get_root function.
This commit will start the process of updating the handling of
the generation of paths for security events that receive either
path or dentry parameters.

The conditionals were re-structured to return directly from the
previous conditional clauses.  This allowed the conditions to be
simplified and the indentation level for the function reduced.