Indicate pathname resolution by get_root function.
The event.c:get_root() function is designed to either set the
name of the device that the filesystem containing the path is
platformed on or to compute the pathname if there is no backing
device.

Update the get_root() function so that the pathname number of the
tsem_path structure is either set to a valid pathname or a NULL
pointer.

Update the calls to the get_root() function so that they return
with a resolved pathname if the pathname pointer is non-NULL or
if the function returns a non-zero error code.