Update handling of non-device based paths.
The event.c:get_root() function is responsible for setting the
device number of a device that is hosting a filesystem or
generating a pathname on a pseudo-filesystem that does not have
device backing.

If the determination is made that a device is not backing the
filesystem, a __getname() based buffer is allocated to handle the
pathname that will be generated.  After the buffer is allocated a
single exit point for the function is used that releases the
buffer allocation.

Previously, due to a design oversight, a pathname was not being
generated for filesystems that did not implement a ->d_dname
dentry method.

Use the dentry_path_raw() function to generate a pathname.  This
change, in particular, causes paths to be generated for the proc
and sysfs filesystems.