Substitute PID values in /proc files.
The implementation of full pathnames for non device backed files
created invariancy problems in the state of security modeling
namespaces due to the fact that pathnames in the /proc filesystem
can be pids.

This commit introduces the events.c:_substitute_pids() function
to replace PID values with constant strings that will result in
invariant state values.

Any path components that match the number PID value of the
process is replaced with a path component of SELF.

This leaves the other issue that requires invariancy attention
are the /proc/PID directories.  The first path component is
checked and if the first path component is a numeric value it is
replaced with a path component of PID.