Substitute PID directories with task identities.
The use of /PID as the name of a directory for a /proc PID directory
is non-specific.  This can be addressed with increased security
resolution by substituting the task identity of the process named by
the PID directory.

Locate the task control structure of the PID directory and write the
hexadecimal expression of the task as identity as the first directory
component, in replace of the PID number.  In the unlikely event that
the task represented by the PID is not running use PID as the first
directory name.