Implement reference description for pathname.
In order to increase the model precision for path descriptions
an additional field was added to the JSON path description that
specifies the context that the pathname was computed in.

The tsem_path_type enumeration was added to support tagging the
path referenced in the tsem_event structure.  The following
enumeration types are implemented which are designed to be
self-explanatory for the pathname reference:

TSEM_PATH_TYPE_ROOT
TSEM_PATH_TYPE_CHROOT
TSEM_PATH_TYPE_NAMESPACE

This enumeration type was added as a member to the tsem_path
structure in order to preserve the path reference.

Since the event.c:get_root() function is called for both path and
dentry based pathnames a multi-component if statement was added
at the top of this function to set the path reference type.

A switch statement was used in both the fs.c:show_path() function
and the map.c:add_path() function to convert the enumeration type
to a string value.

The selected string value is added to the input image for the
generation of the CELL value.

The ABI documentation was updated to include the new key value in
the path description structure.