Handle mapping of temporary files without a pathname.
Testing with the overlayfs implementation indicated that any
attempt to copy a file to the merged filesystem resulted in an
EINVAL error.

The issue was traced to the fact that the overlayfs
implementation references files in a private mount namespace that
cannot generate a full path reference.  As a result the check for
a path separator in the map.c:add_temp_path() function fails
causing an EINVAL error code to be returned for the event.

The logic after the strrchr() function that was used to locate
the trailing path separator was changed to act only if a non-NULL
pointer was returned, ie. a match for a trailing path separator
was found.

If the strrchr() function does not find a trailing slash the
pathname passed to the function is extended into the measurement
rather than returning an EINVAL error code.