Return correct non-error code for security_task_prctl handler.
A regression was noted with the Docker 'hello-world' container
with a TSEM enabled kernel.  When the application terminated the
container hung and would not terminate until a SIGINT signal was
generated by entering a CNTRL-C key combination.

The problem was found to be secondary to the tsem_task_prctl
event handler not returning the correct code to indicate that no
error condition was being generated.

The security_task_prctl event is one of the security events that
does not follow the model of returning 0 to indicate that
permission was permitted.  Instead an event handler is expected
to return an -ENOSYS value to indicate that permission is allowed
for the event.

Check the return code from the model and convert a value of zero
(no-error) into the -ENOSYS value.