Replace process authentication with direct identification.
In early reviews of TSEM Paul Moore had raised concerns regarding
the use of process identifiers (PID's) to locate a task whose
execution status was to be set.

An authentication infrastructure was implemented in order to
address these issues.  This infrastructure was extensively
described and documented but generic and non-specific criticism
was still leveled.

Inspiration was taken from the Task Temporal Disambiguator (TTD)
value, that is used to identify each unique instance of a task
identity, to address this problem.

A 64-bit counter was implemented that is incremented by the
tsem_task_alloc() security event handler for each task that is
allocated.  This value is placed in the tnum member of the
tsem_event structure and serves as a perpetually unique
identifier for each task.

For events that are exported to userspace this identifier is
included along with the PID value.  The control plane calls to
set the trusted or untrusted status of the process were updated
to require the presentation of both the PID and the task number
values.

The syntax of the control plane strings are now as follows:

trusted PID TNUM

untrusted PID TNUM

The PID value is used to locate the candidate process whose
status will be set.  The task number passed in the control plane
call is compared to the task number that was set in the tsem_task
structure when the task was created.  If the values do not match
an error is returned to the trust orchestrator.

All of the infrastructure in the namespace.c file that was used
to manage the process identification keys was dropped.

This strategy eliminates any possibility of an attack that
focuses on wrapping a PID value since the task number will not
wrap in any conceivable system lifetime.  If a process were to be
forked every nanosecond it would take 54 years for the task
identifier value to wrap.

The TSEM documentation was updated to remove the discussion of
the authentication infrastructure.  A descriptiption of the task
identifier was used to replace that information.

A discussion in the documentation was also included as to why,
even is possible, why a process substitution task would be
ineffective, secondary to the task identity model that TSEM is
based on.

The ABI documentation was also updated to reflect the new syntax
for the control plane values.  The event description was updated
to include documentation of the pid and tnum key/value pairs with
an explanation that these values are only included when an event
description is exported to userspace.