Escape quotation in paths.
According to the JSON specification a double quote " must be
escaped by prefixing it with a backslash when the quote appears
in a JSON string.

For simplicity all of the values for the key/vale pairs of the
JSON event coding are string values so a filename that would
happen to contain a double quote would need to have quote
'escaped'.

The _escape_path() private function was added that implements the
replacement of any double quotes in the pathname with the
'escaped' version, ie. with a backslash prepended.  Each path
that is produced is passed through this function as the final
step in the preparation of the path description.