Set the task number at task allocation time.
A regression was noted with externally modeled namespaces where
the task number in the exported description was zero, which
should be an issue that is impossible.  The events with a zero
task number spawned by runc and were believed to be threads of
the primary task.

The situation was addressed by setting the task number of the new
task to the task number of the process/task that was creating the
new task.

This addresses the problem and produces export descriptions that
are properly orchestrated.