Do not reference arg for security_ioctl handlers.
The argument value for the following security event handlers:

tsem_file_ioctl

tsem_file_ioctl_compat

Can be either a numeric file or a memory pointer.

The security event handlers are not allowed to access the memory
pointer if the argument represents a userspace pointer value.

Given the uncertainty in origin of this value it is not handled
by any of the TSEM infrastructure and it would be unsafe to allow
any custom models to reference the value.

Remove the capture of this value in both event handlers so the
argument is neither persisted or exported.