Implement direct namespace access for orchestrator process.
This commit begins implementation of a major change in the
security modeling namespace management architecture.

When a security modeling namespace is created a pointer is placed
in the tsem_task control structure for the process that will
serve as the trust orchestrator for the security modeling
namespace.  This allows the trust orchestrator to access the
control plane file that is exporting security events from an
externally modeled namespace or the the model characteristic of
an internally modeled namespace.

This provides kernel support for simplified userspace trust
orchestrators.  Since the process that spans the lead process in
a security modeling namespace has access to the security modeling
context of the namespace there is no need to have a monitor
process running in the namespace that exports events or model
characteristics to the orchestrator process.

In order to support this new model the tsem_task structure had
the tma_for_ns member replaced with a tsem_context structure is
named tma_context.  This structure points to security modeling
context that is maintained by the security modeling namespace.

The tsem_external structure that encapsulates the information
needed for the export of security events was updated to replace
the have_event boolean member with an unsigned integer,
event_cnt, that carries the the number of outstanding events that
need to be exported.

The tsem_tma_context helper() function was added to encapsulate
the return of the security modeling context.

The fs.c:control_COE() function was updated to verify that the
security modeling context of the process that is being acted on
matches the context that is held by the process issueing the
orchestration commands.

In order to address issues that were observer with the userspace
orchestration utilities a count of the number of outstanding
events to be exported is now maintained in the event_cnt member
of the tsem_external structure.  That value is incremented inside
of the spinlock that protects the list of outstanding events when
an event is registered for export.  Conversely the value is
decremented by the tsem_export_show() function when a request is
made to read the event queu.

The fs.c:export_poll() function was updated to check the count of
outstanding events when a poll call on the control plane file.
A POLLIN condition is raised if there are any outstanding events
to be exported.