Select the proper security modeling context for exported events.
In order to support the export of security events from the root
modeling namespace the tsem_tma_context() function had to be
updated to check the context of the task that is passed to the
function in order to select the correct context to return.

If the task is from the root modeling namespace the context
describing the root modeling namespace is returned.  If the
exports are from a subordinate modeling namespace the context
held by the trust orchestrator is returned.