Properly reference control plane.
Testing with internally modeled namespaces indicated that the
trust orchestrator for these namespaces was not reference the
correct namespace context.

The definition tsem_tma_context() was returned to its original
implementation that returns the tsem_context structure for the
task structure that it was supplied to.  It no longer attempts to
make any corrections based on whether this is a root or secondary
context.

In the tsem_export_show() function a conditional was added at the
top of the function to determine if the task that was exporting
the event was in the root modeling namespace.  If so context that
was supplied by tsem_tma_context() was replaced with the
namespace context of the current task.

The following inline function was introduced in the fs.c file to
return a pointer to the model that is being implemented for a
specific task:

tma_model()

All of the calls to the control plane that return characteristics
of a model were modified to use this this call in order to access
the model definition for the calling process, that will now
always be the trust orchestrator process.